claude-code-observability-plugin/tests/requirements.txt
pytest>=7.0
pyyaml>=6.0
pydantic>=2.0
requests>=2.28
hypothesis>=6.0
